Added UpDown into spinal.lib.bus.fabric
