+++ lib/src/main/scala/spinal/lib/bus/tilelink/fabric/Connection.scala
@@ -12,7 +12,7 @@ import scala.collection.mutable.ArrayBuffer
 /**
  * Implementation of ConnectionBase which allows the automatic insertion of bridges
  */
-class Connection(m : NodeBase, s : NodeBase) extends ConnectionBase(m, s) {
+class Connection(m : NodeUpDown, s : NodeUpDown) extends ConnectionBase(m, s) {
 
   //Will negociate the parameters and then connect the ends through the required adapters
   val thread = Fiber build new Area{
